% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{APA-formatted p-values}
\usage{
format_p(p, digits = 3, minp = 0.001)
}
\arguments{
\item{p}{the existing p-value (should be numeric, >0, <1)}

\item{digits}{integer indicating the number of decimal places to be used for rounding and display (1:10)}

\item{minp}{the minimum pvalue under which the return value defaults to p < pmin}
}
\value{
character
}
\description{
\code{format_p} Produces an APA-formatted string for reporting a p-value.
}
\examples{
format_p(0.03007) # p = .030
format_p(0.34567, digits = 4) # p = .3457
format_p(0.005, digits = 2, minp = .01) # p < .01
}
