% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{r2z}
\alias{r2z}
\title{Fisher's R to Z}
\usage{
r2z(r)
}
\arguments{
\item{r}{the existing r-value (should be numeric, >-1, <1)}
}
\value{
double
}
\description{
\code{r2z} Converts an r-value to a z-value using Fisher's r-to-z transformation.
}
\examples{
r2z(-.4) # -0.4236489
}
