% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{steiger_dat}
\alias{steiger_dat}
\title{Steiger's Z for Data Frames or vectors}
\usage{
steiger_dat(v1, v2 = NULL, v3 = NULL, alpha = 0.05, tails = 2)
}
\arguments{
\item{v1}{the first vector or a dataframe where the first three columns are the vectors}

\item{v2}{the second vector (ignored if v1 is a data frame)}

\item{v3}{the third vector (ignored if v1 is a data frame)}

\item{alpha}{the number of observations}

\item{tails}{whether to report a one-tailed or two-tailed p-value}
}
\value{
list
}
\description{
\code{steiger} Test for significant difference between two correlations on the same sample.
}
\examples{
steiger_dat(iris)
}
